added : Variant::operator==(int), Variant parses preceding space

+++ tests/tObject.cpp
@@ -9,7 +9,7 @@ TEST(tObject, Parse) {
     EXPECT_TRUE(a.contains("a"));
     EXPECT_FALSE(a.contains("b"));
     EXPECT_EQ(a.size(), 1);
-    EXPECT_EQ(static_cast<double>(a["a"]), 4);
+    EXPECT_EQ(a["a"], 4);
 }
 
 TEST(tObject, Empty) {
@@ -19,7 +19,7 @@ TEST(tObject, Empty) {
 }
 TEST(tObject, TerminatingCommas) {
     Object b("{\"a\":4.1,\"b\":false,}");
-    EXPECT_EQ(static_cast<double>(b["a"]), 4.1);
+    EXPECT_EQ(b["a"], 4.1);
     EXPECT_EQ(b["b"], false);
     EXPECT_EQ(b.size(), 2);
 }
